requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
numpy==2.1.1
packaging==24.1
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
Telethon==1.36.0
tzdata==2024.1
uritemplate==4.1.1
